refactor(gamestate): split event-handler regions into GameStateRecorder.Handlers.cs
