+++ Source/Parsek/GameStateRecorder.cs
@@ -11,7 +11,7 @@ namespace Parsek
     /// Records career-mode game state changes (contracts, tech, crew, facilities, resources).
     /// Lifecycle managed by ParsekScenario — Subscribe/Unsubscribe called on every OnLoad.
     /// </summary>
-    internal class GameStateRecorder
+    internal partial class GameStateRecorder
     {
         /// <summary>
         /// Set to true by ParsekScenario during its own crew mutations
@@ -387,612 +387,6 @@ namespace Parsek
 
         #endregion
 
-        #region Contract Handlers
-
-        private void OnContractOffered(Contract contract)
-        {
-            if (contract == null)
-            {
-                LogEventReject("ContractOffered", "contract-null");
-                return;
-            }
-            var title = contract.Title ?? "";
-            // #398: deliberately do NOT add to GameStateStore. Offered contracts are
-            // transient advertisements generated by KSP's ContractSystem tick — they are
-            // not a player action, and every generation cycle creates fresh GUIDs that
-            // never dedup. The converter already drops ContractOffered, and no UI or
-            // module reads Offered events from the store. Keep the subscription for the
-            // diagnostic log only.
-            ParsekLog.Verbose("GameStateRecorder",
-                $"Game state: ContractOffered '{title}' (diagnostic, not stored)");
-        }
-
-        private void OnContractAccepted(Contract contract)
-        {
-            if (contract == null)
-            {
-                LogEventReject("ContractAccepted", "contract-null");
-                return;
-            }
-            string guid = contract.ContractGuid.ToString();
-
-            var title = contract.Title ?? "";
-            double deadline = contract.TimeDeadline;
-            float advanceFunds = (float)contract.FundsAdvance;
-            float failFunds = (float)contract.FundsFailure;
-            float failRep = (float)contract.ReputationFailure;
-            string contractType = contract.GetType().Name ?? "";
-            ConfigNode contractNode = null;
-            Exception snapshotFailure = null;
-
-            try
-            {
-                contractNode = new ConfigNode("CONTRACT");
-                contract.Save(contractNode);
-                string savedContractType = contractNode.GetValue("type");
-                if (!string.IsNullOrEmpty(savedContractType))
-                    contractType = savedContractType;
-            }
-            catch (Exception ex)
-            {
-                snapshotFailure = ex;
-                contractNode = null;
-            }
-
-            // Structured detail: title + deadline + type + advance + failure penalties.
-            // advance must be captured at accept time — KSP applies it immediately via
-            // FundsChanged(ContractAdvance), which the converter intentionally drops.
-            // Without funds= here, AdvanceFunds stayed 0 and FundsModule never credited
-            // the advance (codex review [P1] on PR #307). type= must also be captured
-            // here because ContractAccept ledger actions are consumed by UI/policy paths
-            // that do not re-open the stored snapshot.
-            // deadline=0 means no deadline (KSP convention) — store as NaN.
-            string deadlineStr = deadline > 0
-                ? deadline.ToString("R", System.Globalization.CultureInfo.InvariantCulture)
-                : "NaN";
-            var detail = $"title={title};deadline={deadlineStr}" +
-                (string.IsNullOrEmpty(contractType) ? "" : $";type={contractType}") +
-                $";funds={advanceFunds.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}" +
-                $";failFunds={failFunds.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}" +
-                $";failRep={failRep.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}";
-
-            var evt = new GameStateEvent
-            {
-                ut = Planetarium.GetUniversalTime(),
-                eventType = GameStateEventType.ContractAccepted,
-                key = guid,
-                detail = detail
-            };
-            Emit(ref evt, "ContractAccepted");
-
-            // Store full contract snapshot for reversal
-            if (contractNode != null)
-            {
-                GameStateStore.AddContractSnapshot(guid, contractNode);
-                ParsekLog.Info("GameStateRecorder",
-                    $"Game state: ContractAccepted '{title}' type='{contractType}' deadline={deadlineStr} " +
-                    $"advance={advanceFunds} failFunds={failFunds} failRep={failRep} (snapshot saved)");
-            }
-            else
-            {
-                ParsekLog.Warn("GameStateRecorder",
-                    $"Game state: ContractAccepted '{title}' type='{contractType}' " +
-                    $"(snapshot FAILED: {(snapshotFailure != null ? snapshotFailure.Message : "unknown error")})");
-            }
-
-            // Write directly to ledger when this event has no recording owner and no
-            // live recorder can still claim it. Tagged contract events flow through
-            // the normal commit-time ConvertEvents path. Untagged pre-recording
-            // FLIGHT events must be direct-ledger too: stock can complete launch-site
-            // contracts just before Parsek auto-record starts, before a live recorder
-            // exists.
-            // #405: without this, accepted contracts at KSC never reached the ledger and
-            // PatchContracts had no active contracts to preserve on next recalc.
-            // #431 gate: during FLIGHT -> SPACECENTER teardown the scene already reads
-            // SPACECENTER but ParsekFlight.Instance is still alive and Emit legitimately
-            // tagged this event with the outgoing recordingId. Skip the ledger write
-            // when non-empty: the tagged event is already in
-            // GameStateStore and will flow through the commit-time path (or be purged
-            // with the recording on discard). Without this gate the ledger would end up
-            // with an untagged KSC action that survives DiscardPendingTree forever.
-            // If the tag is empty while a live recorder exists, treat it as tag drift
-            // and avoid creating a null-owner ledger action that survives discard.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        private void OnContractCompleted(Contract contract)
-        {
-            if (contract == null)
-            {
-                LogEventReject("ContractCompleted", "contract-null");
-                return;
-            }
-
-            // Bug 1 (fix-funds-economy-divergence §3.2): debounce a stock contract re-fire.
-            // Stock KSP can fire onCompleted N times for one contract within a few ms (the
-            // observed burst fired 4x at the same UT). RETURN before BOTH the Emit store
-            // write AND the ShouldForwardDirectLedgerEvent -> OnKscSpending forward below,
-            // so neither inflation path (tagged-commit via the store, no-owner-direct via
-            // OnKscSpending) bakes a duplicate reward. The window is the same coalesce window
-            // the carried resource events already collapse within.
-            string contractGuid = contract.ContractGuid.ToString();
-            double completionUt = Planetarium.GetUniversalTime();
-            if (IsDuplicateContractCompletion(
-                contractGuid, completionUt,
-                lastContractCompletionUtByGuid, GameStateStore.ResourceCoalesceEpsilon))
-            {
-                ParsekLog.Info("GameStateRecorder",
-                    $"Game state: ContractCompleted '{contract.Title ?? ""}' DEDUPED " +
-                    $"(stock re-fire of contract {contractGuid} within {GameStateStore.ResourceCoalesceEpsilon.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}s at UT={completionUt.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}) - not recorded");
-                return;
-            }
-
-            var title = contract.Title ?? "";
-            // InvariantCulture-safe float serialization — plain interpolation emits
-            // comma decimals on locales like de-DE/fr-FR/ro-RO, and the IC-parsing
-            // converter side then reads zero. See also OnContractAccepted.
-            float fundsReward = (float)contract.FundsCompletion;
-            float repReward = (float)contract.ReputationCompletion;
-            float sciReward = (float)contract.ScienceCompletion;
-            var detail = $"title={title}" +
-                $";fundsReward={fundsReward.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}" +
-                $";repReward={repReward.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}" +
-                $";sciReward={sciReward.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}";
-            var evt = new GameStateEvent
-            {
-                ut = completionUt,
-                eventType = GameStateEventType.ContractCompleted,
-                key = contractGuid,
-                detail = detail
-            };
-            Emit(ref evt, "ContractCompleted");
-            ParsekLog.Info("GameStateRecorder",
-                $"Game state: ContractCompleted '{title}' (funds={fundsReward}, rep={repReward}, sci={sciReward})");
-
-            // #405: route to ledger immediately when there is no recording owner.
-            // #431 gate: see OnContractAccepted — skip the ledger write when the event
-            // was tagged during FLIGHT -> SPACECENTER teardown.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        private void OnContractFailed(Contract contract)
-        {
-            if (contract == null)
-            {
-                LogEventReject("ContractFailed", "contract-null");
-                return;
-            }
-            var title = contract.Title ?? "";
-            float fundsPenalty = (float)contract.FundsFailure;
-            float repPenalty = (float)contract.ReputationFailure;
-            var detail = $"title={title}" +
-                $";fundsPenalty={fundsPenalty.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}" +
-                $";repPenalty={repPenalty.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}";
-            var evt = new GameStateEvent
-            {
-                ut = Planetarium.GetUniversalTime(),
-                eventType = GameStateEventType.ContractFailed,
-                key = contract.ContractGuid.ToString(),
-                detail = detail
-            };
-            Emit(ref evt, "ContractFailed");
-            ParsekLog.Info("GameStateRecorder",
-                $"Game state: ContractFailed '{title}' (fundsPenalty={fundsPenalty}, repPenalty={repPenalty})");
-
-            // #405: route to ledger immediately when there is no recording owner.
-            // #431 gate: see OnContractAccepted.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        private void OnContractCancelled(Contract contract)
-        {
-            if (contract == null)
-            {
-                LogEventReject("ContractCancelled", "contract-null");
-                return;
-            }
-            var title = contract.Title ?? "";
-            float fundsPenalty = (float)contract.FundsFailure;
-            float repPenalty = (float)contract.ReputationFailure;
-            var detail = $"title={title}" +
-                $";fundsPenalty={fundsPenalty.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}" +
-                $";repPenalty={repPenalty.ToString("R", System.Globalization.CultureInfo.InvariantCulture)}";
-            var evt = new GameStateEvent
-            {
-                ut = Planetarium.GetUniversalTime(),
-                eventType = GameStateEventType.ContractCancelled,
-                key = contract.ContractGuid.ToString(),
-                detail = detail
-            };
-            Emit(ref evt, "ContractCancelled");
-            ParsekLog.Info("GameStateRecorder",
-                $"Game state: ContractCancelled '{title}' (fundsPenalty={fundsPenalty}, repPenalty={repPenalty})");
-
-            // #405: route to ledger immediately when there is no recording owner.
-            // #431 gate: see OnContractAccepted.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        private void OnContractDeclined(Contract contract)
-        {
-            if (contract == null)
-            {
-                LogEventReject("ContractDeclined", "contract-null");
-                return;
-            }
-            var title = contract.Title ?? "";
-            var evt = new GameStateEvent
-            {
-                ut = Planetarium.GetUniversalTime(),
-                eventType = GameStateEventType.ContractDeclined,
-                key = contract.ContractGuid.ToString(),
-                detail = title
-            };
-            Emit(ref evt, "ContractDeclined");
-            ParsekLog.Info("GameStateRecorder", $"Game state: ContractDeclined '{title}'");
-        }
-
-        #endregion
-
-        #region Tech Handlers
-
-        private void OnTechResearched(GameEvents.HostTargetAction<RDTech, RDTech.OperationResult> data)
-        {
-            if (IsReplayingActions)
-            {
-                ParsekLog.Verbose("GameStateRecorder", "Suppressed TechResearched event during action replay");
-                return;
-            }
-            if (data.host == null)
-            {
-                LogEventReject("TechResearched", "tech-null");
-                return;
-            }
-            if (data.target != RDTech.OperationResult.Successful)
-            {
-                LogEventReject(
-                    "TechResearched",
-                    "operation-not-successful",
-                    data.host.techID,
-                    data.target.ToString());
-                return;
-            }
-
-            string techId = data.host.techID ?? "";
-            string partList = "";
-            if (data.host.partsAssigned != null && data.host.partsAssigned.Count > 0)
-            {
-                var names = new List<string>();
-                foreach (var part in data.host.partsAssigned)
-                {
-                    if (part != null)
-                        names.Add(part.name ?? "");
-                }
-                partList = string.Join(",", names);
-            }
-
-            var evt = new GameStateEvent
-            {
-                ut = Planetarium.GetUniversalTime(),
-                eventType = GameStateEventType.TechResearched,
-                key = techId,
-                detail = $"cost={data.host.scienceCost};parts={partList}",
-                valueBefore = ResearchAndDevelopment.Instance != null
-                    ? ResearchAndDevelopment.Instance.Science + data.host.scienceCost
-                    : 0,
-                valueAfter = ResearchAndDevelopment.Instance != null
-                    ? ResearchAndDevelopment.Instance.Science
-                    : 0
-            };
-            Emit(ref evt, "TechResearched");
-            ParsekLog.Info("GameStateRecorder", $"Game state: TechResearched '{techId}' (cost={data.host.scienceCost})");
-
-            // Write directly to ledger when there is no recording owner.
-            // #553 follow-up: gate on the same ShouldForwardDirectLedgerEvent predicate
-            // used by the contract handlers. Untagged pre-recording FLIGHT tech-research
-            // events have no later commit-time owner; suppressing them here would
-            // strand the TechResearched action in GameStateStore.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        private void OnPartPurchased(AvailablePart part)
-        {
-            if (IsReplayingActions)
-            {
-                ParsekLog.Verbose("GameStateRecorder", "Suppressed PartPurchased event during action replay");
-                return;
-            }
-            if (part == null)
-            {
-                LogEventReject("PartPurchased", "part-null");
-                return;
-            }
-            var partName = part.name ?? "";
-            // InvariantCulture-safe: plain interpolation serializes floats with the
-            // system locale decimal separator, and ConvertPartPurchased parses with IC.
-            var ic = System.Globalization.CultureInfo.InvariantCulture;
-            float entryCost = part.entryCost;
-            float chargedCost = ComputePartPurchaseFundsSpent(entryCost);
-
-            // #451: `cost=` stays authoritative for save/load and UI because it must
-            // reflect the ACTUAL funds debit: 0 when BypassEntryPurchaseAfterResearch is
-            // on, entryCost when the harder no-bypass difficulty is active. Persist the
-            // raw stock entry price separately for diagnostics and future tooling.
-            var detail = "cost=" + chargedCost.ToString("R", ic) +
-                         ";entryCost=" + entryCost.ToString("R", ic);
-
-            // Keep before/after in sync with the charged amount so future consumers that
-            // derive deltas from the numeric fields do not resurrect the bypass=true bug.
-            var evt = new GameStateEvent
-            {
-                ut = Planetarium.GetUniversalTime(),
-                eventType = GameStateEventType.PartPurchased,
-                key = partName,
-                detail = detail,
-                valueBefore = Funding.Instance != null ? Funding.Instance.Funds + chargedCost : 0,
-                valueAfter = Funding.Instance != null ? Funding.Instance.Funds : 0
-            };
-            Emit(ref evt, "PartPurchased");
-            ParsekLog.Info("GameStateRecorder",
-                $"Game state: PartPurchased '{partName}' (chargedCost={chargedCost}, entryCost={entryCost})");
-
-            // #405: route to ledger immediately when there is no recording owner.
-            // Relies on the DedupKey (§F) to disambiguate part-name collisions.
-            // #553 follow-up: gate on ShouldForwardDirectLedgerEvent so untagged
-            // pre-recording FLIGHT part-purchases reach the ledger too.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        /// <summary>
-        /// Returns the actual funds debit for an R&D part purchase. Stock KSP still
-        /// fires OnPartPurchased when BypassEntryPurchaseAfterResearch is enabled, but
-        /// the purchase is free in that mode.
-        /// </summary>
-        internal static float ComputePartPurchaseFundsSpent(float entryCost)
-        {
-            return IsBypassEntryPurchaseAfterResearch() ? 0f : entryCost;
-        }
-
-        #endregion
-
-        #region Crew Handlers
-
-        /// <summary>
-        /// Fires from KerbalRoster.HireApplicant when the player pays to recruit an
-        /// applicant from the Astronaut Complex. <paramref name="activeCrewCount"/> is
-        /// KSP's pre-hire crew count — the same value GetRecruitHireCost takes to pick
-        /// the cost curve tier.
-        /// </summary>
-        private void OnCrewmemberHired(ProtoCrewMember crew, int activeCrewCount)
-        {
-            if (IsReplayingActions)
-            {
-                ParsekLog.Verbose("GameStateRecorder", "Suppressed CrewHired event during action replay");
-                return;
-            }
-            if (SuppressCrewEvents)
-            {
-                ParsekLog.VerboseRateLimited("GameStateRecorder", "suppress-crew-hired",
-                    "Suppressed CrewHired event while Parsek mutates crew state", 5.0);
-                return;
-            }
-            if (crew == null)
-            {
-                LogEventReject("CrewHired", "crew-null");
-                return;
-            }
-            var name = crew.name ?? "";
-
-            // Hire cost from KSP's career curve. activeCrewCount is passed in by
-            // KerbalRoster; prefer it over re-reading the roster to avoid a race with
-            // the upcoming type flip (Applicant -> Crew) inside HireApplicant.
-            float hireCost = ComputeHireCost(activeCrewCount);
-
-            // InvariantCulture-safe hire cost — GetRecruitHireCost is typically integer
-            // but could return fractional on modded career curves.
-            var ic = System.Globalization.CultureInfo.InvariantCulture;
-            var evt = new GameStateEvent
-            {
-                ut = Planetarium.GetUniversalTime(),
-                eventType = GameStateEventType.CrewHired,
-                key = name,
-                detail = $"trait={crew.trait ?? ""};cost={hireCost.ToString("R", ic)}"
-            };
-            Emit(ref evt, "CrewHired");
-            ParsekLog.Info("GameStateRecorder",
-                $"Game state: CrewHired '{name}' ({crew.trait ?? "?"}) " +
-                $"cost={hireCost.ToString("R", ic)} activeCrewCount={activeCrewCount}");
-
-            // #553 follow-up: gate on ShouldForwardDirectLedgerEvent so untagged
-            // pre-recording FLIGHT crew hires reach the ledger too.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        /// <summary>
-        /// Returns the KSP hire cost for the Nth recruit. Safe to call in tests or when
-        /// GameVariables is not yet initialized — returns 0f in that case.
-        /// </summary>
-        internal static float ComputeHireCost(int activeCrewCount)
-        {
-            if (GameVariables.Instance == null || HighLogic.CurrentGame == null)
-                return 0f;
-            return (float)GameVariables.Instance.GetRecruitHireCost(activeCrewCount);
-        }
-
-        private void OnKerbalRemoved(ProtoCrewMember crew)
-        {
-            if (SuppressCrewEvents)
-            {
-                ParsekLog.VerboseRateLimited("GameStateRecorder", "suppress-crew-removed",
-                    "Suppressed CrewRemoved event while Parsek mutates crew state", 5.0);
-                return;
-            }
-            if (crew == null)
-            {
-                LogEventReject("CrewRemoved", "crew-null");
-                return;
-            }
-            var name = crew.name ?? "";
-
-            var evt = new GameStateEvent
-            {
-                ut = Planetarium.GetUniversalTime(),
-                eventType = GameStateEventType.CrewRemoved,
-                key = name,
-                detail = $"trait={crew.trait ?? ""}"
-            };
-            Emit(ref evt, "KerbalRemoved");
-            ParsekLog.Info("GameStateRecorder", $"Game state: CrewRemoved '{name}'");
-        }
-
-        /// <summary>
-        /// Returns true if we're currently in the Flight scene. KSC spending actions
-        /// should only be written directly to the ledger outside of flight (to avoid
-        /// double-adding when the recording is committed).
-        /// </summary>
-        private static bool IsFlightScene()
-        {
-            return HighLogic.LoadedScene == GameScenes.FLIGHT;
-        }
-
-        /// <summary>
-        /// Returns true when a captured lifecycle event has no recording owner and
-        /// should be written straight to the ledger. A live recorder with an empty
-        /// tag is tag drift, not proof that the event is ownerless; true
-        /// pre-recording FLIGHT events have no tag and no live recorder yet.
-        /// </summary>
-        internal static bool ShouldForwardDirectLedgerEvent(string recordingTag, bool hasLiveRecorder)
-        {
-            return string.IsNullOrEmpty(recordingTag) && !hasLiveRecorder;
-        }
-
-        internal static bool ShouldForwardDirectScienceSubject(
-            string recordingTag,
-            bool hasLiveRecorder,
-            bool hasActiveUncommittedTree)
-        {
-            return ShouldForwardDirectLedgerEvent(recordingTag, hasLiveRecorder) &&
-                   !hasActiveUncommittedTree;
-        }
-
-        /// <summary>
-        /// Pure: returns true if the status transition is a real change (not an identity
-        /// transition like Dead->Dead). Extracted for testability (Bug #122).
-        /// </summary>
-        internal static bool IsRealStatusChange(ProtoCrewMember.RosterStatus oldStatus,
-            ProtoCrewMember.RosterStatus newStatus)
-        {
-            return oldStatus != newStatus;
-        }
-
-        /// <summary>
-        /// Returns true if a crew status change event should be suppressed.
-        /// Suppresses when: (1) bulk mutation in progress, (2) crew is managed by Parsek
-        /// (reserved/stand-in — KSP oscillates their status as noise), (3) identity
-        /// transition (same status, e.g. Dead→Dead).
-        /// Pure static for testability — no KSP state access.
-        /// </summary>
-        internal static bool ShouldSuppressCrewStatusChange(
-            string crewName, bool suppressFlag, bool isIdentity)
-        {
-            if (suppressFlag) return true;
-            if (crewName != null && (LedgerOrchestrator.Kerbals?.IsManaged(crewName) ?? false)) return true;
-            if (isIdentity) return true;
-            return false;
-        }
-
-        private void OnKerbalStatusChange(ProtoCrewMember crew,
-            ProtoCrewMember.RosterStatus oldStatus, ProtoCrewMember.RosterStatus newStatus)
-        {
-            if (crew == null)
-            {
-                LogEventReject("CrewStatusChanged", "crew-null");
-                return;
-            }
-
-            bool isIdentity = !IsRealStatusChange(oldStatus, newStatus);
-            if (ShouldSuppressCrewStatusChange(crew.name, SuppressCrewEvents, isIdentity))
-            {
-                if (SuppressCrewEvents)
-                    ParsekLog.VerboseRateLimited("GameStateRecorder", "suppress-crew-status",
-                        "Suppressed CrewStatusChanged event while Parsek mutates crew state", 5.0);
-                else if (isIdentity)
-                    ParsekLog.Verbose("GameStateRecorder",
-                        $"Filtered identity crew status transition for '{crew.name}': {oldStatus} -> {newStatus}");
-                else
-                    ParsekLog.VerboseRateLimited("GameStateRecorder", "suppress-managed-crew",
-                        $"Suppressed CrewStatusChanged for managed kerbal '{crew.name}': {oldStatus} -> {newStatus}", 5.0);
-                return;
-            }
-
-            var name = crew.name ?? "";
-
-            double ut = Planetarium.GetUniversalTime();
-
-            // Debounce: if we have a pending event for this crew within 0.1s game time
-            // and this event reverses the previous one, cancel both.
-            PendingCrewEvent prev;
-            if (pendingCrewEvents.TryGetValue(name, out prev) && Math.Abs(prev.gameEvent.ut - ut) < 0.1)
-            {
-                if (prev.to == oldStatus && newStatus == prev.from)
-                {
-                    GameStateStore.RemoveEvent(prev.gameEvent);
-                    pendingCrewEvents.Remove(name);
-                    ParsekLog.Verbose("GameStateRecorder", $"Debounced crew status bounce for '{name}'");
-                    return;
-                }
-            }
-
-            var evt = new GameStateEvent
-            {
-                ut = ut,
-                eventType = GameStateEventType.CrewStatusChanged,
-                key = name,
-                detail = $"from={oldStatus};to={newStatus}"
-            };
-            Emit(ref evt, "KerbalStatusChange");
-            pendingCrewEvents[name] = new PendingCrewEvent { gameEvent = evt, from = oldStatus, to = newStatus };
-            ParsekLog.Info("GameStateRecorder", $"Game state: CrewStatusChanged '{name}' {oldStatus} → {newStatus}");
-        }
-
-        private void OnKerbalTypeChange(ProtoCrewMember pcm, ProtoCrewMember.KerbalType fromType, ProtoCrewMember.KerbalType toType)
-        {
-            if (SuppressCrewEvents) return;
-
-            if (pcm == null)
-            {
-                LogEventReject("KerbalTypeChanged", "crew-null");
-                return;
-            }
-
-            // Only care about Unowned -> Crew transitions (rescue pickup)
-            if (fromType != ProtoCrewMember.KerbalType.Unowned || toType != ProtoCrewMember.KerbalType.Crew)
-                return;
-
-            string name = pcm?.name ?? "";
-            string trait = pcm?.experienceTrait?.TypeName ?? "Pilot";
-
-            var evt = new GameStateEvent
-            {
-                ut = Planetarium.GetUniversalTime(),
-                eventType = GameStateEventType.KerbalRescued,
-                key = name,
-                detail = $"trait={trait}"
-            };
-            Emit(ref evt, "KerbalRescued");
-
-            ParsekLog.Info("GameStateRecorder", $"Game state: KerbalRescued '{name}' (trait={trait})");
-        }
-
-        #endregion
-
         #region Resource Handlers
 
         private void SeedResourceState()
@@ -1329,711 +723,6 @@ namespace Parsek
                 $"Science subject captured: {subject.id} amount={amount:F1} total={subject.science:F1} " +
                 $"reason='{reasonKey}' ut={captureUt:F1} tag='{subjectRecordingId}' " +
                 $"directLedger={directLedgerHandled}");
-        }
-
-        #endregion
-
-        #region Progress Milestone Handlers
-
-        private void OnProgressComplete(ProgressNode node)
-        {
-            if (IsReplayingActions)
-            {
-                ParsekLog.Verbose("GameStateRecorder", "Suppressed MilestoneAchieved event during action replay");
-                return;
-            }
-            if (node == null)
-            {
-                ParsekLog.Verbose("GameStateRecorder", "OnProgressComplete: null node — skipped");
-                return;
-            }
-
-            string milestoneId = QualifyMilestoneId(node);
-            if (string.IsNullOrEmpty(milestoneId))
-            {
-                ParsekLog.Verbose("GameStateRecorder", "OnProgressComplete: empty node Id — skipped");
-                return;
-            }
-
-            // Route to the testable helper with a live UT from Planetarium. Production
-            // is the only call site that touches Unity statics; tests call
-            // RegisterPendingMilestoneEvent directly with a literal UT.
-            RegisterPendingMilestoneEvent(milestoneId, Planetarium.GetUniversalTime());
-        }
-
-        /// <summary>
-        /// #443: emit-and-register half of <see cref="OnProgressComplete"/>, extracted as
-        /// an internal static so the cached pending-map entry vs. stored-slot identity
-        /// match is directly testable without going through the live
-        /// <see cref="Planetarium"/>/<see cref="GameEvents"/> plumbing.
-        ///
-        /// Internal static for testability.
-        /// </summary>
-        internal static void RegisterPendingMilestoneEvent(string milestoneId, double ut)
-        {
-            // #400: KSP's reward pipeline is:
-            //   subclass handler -> ProgressNode.Complete() -> OnProgressComplete.Fire()
-            //                    -> [this handler runs here, rewards NOT yet applied]
-            //                    -> subclass calls AwardProgressStandard() -> AwardProgress(..funds, sci, rep..)
-            //
-            // Because rewards land AFTER this subscriber fires, we emit the event with
-            // zeros as defaults and rely on the AwardProgressPatch Harmony postfix to
-            // update the detail in place once it has the real reward values. For non-
-            // career modes the zeros are correct (no rewards apply).
-            var evt = new GameStateEvent
-            {
-                ut = ut,
-                eventType = GameStateEventType.MilestoneAchieved,
-                key = milestoneId,
-                detail = BuildMilestoneDetail(0.0, 0f, 0.0)
-            };
-            Emit(ref evt, "MilestoneAchieved");
-
-            // Emit/AddEvent take `ref GameStateEvent`, so `evt` here already carries the
-            // final identity fields — no mirror needed. The cached copy below is
-            // guaranteed to match the stored slot.
-            //
-            // #443: Tag node.Id -> event association (was ProgressNode reference pre-#443).
-            // The key is the qualified milestone id (e.g. "Kerbin/Landing"), which both
-            // OnProgressComplete and the AwardProgress postfix derive deterministically
-            // via QualifyMilestoneId(node). Re-keying by string removes any future
-            // exposure to ProgressNode instance-aliasing should KSP ever rebuild nodes
-            // mid-Complete().
-            PendingMilestoneEventById[milestoneId] = evt;
-
-            ParsekLog.Info("GameStateRecorder",
-                $"Game state: MilestoneAchieved '{milestoneId}' (awaiting reward enrichment)");
-
-            // Milestones can fire at KSC (e.g., facility-related) or in flight.
-            // Write directly to ledger when no recording owner exists.
-            // #553 follow-up: gate on ShouldForwardDirectLedgerEvent so untagged
-            // pre-recording FLIGHT milestones reach the ledger too.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        /// <summary>
-        /// Formats milestone reward values into the detail string format that
-        /// <see cref="GameStateEventConverter.ConvertMilestoneAchieved"/> parses.
-        /// Internal static for testability.
-        /// </summary>
-        internal static string BuildMilestoneDetail(double funds, float rep, double sci)
-        {
-            var ic = System.Globalization.CultureInfo.InvariantCulture;
-            return $"funds={funds.ToString("R", ic)};" +
-                   $"rep={rep.ToString("R", ic)};" +
-                   $"sci={sci.ToString("R", ic)}";
-        }
-
-        /// <summary>
-        /// Short-lived map of qualified milestone id -> pending MilestoneAchieved event.
-        /// Populated by <see cref="OnProgressComplete"/> when the event is emitted with
-        /// zero rewards, and read by <see cref="EnrichPendingMilestoneRewards"/> (called
-        /// from the Harmony postfix on <c>ProgressNode.AwardProgress</c>) to update the
-        /// detail string in place with the real funds/rep/sci values from the award
-        /// parameters.
-        ///
-        /// #443: Re-keyed from <c>ProgressNode</c> reference to qualified id string. The
-        /// id is derived deterministically by <see cref="QualifyMilestoneId"/> at both the
-        /// emit and enrich call sites, so the lookup no longer depends on KSP handing the
-        /// same instance to both code paths. The cached <see cref="GameStateEvent"/> value
-        /// keeps the same identity fields as the stored slot, so the later
-        /// <see cref="GameStateStore.UpdateEventDetail"/> call can patch the correct row
-        /// without any epoch mirror.
-        ///
-        /// Entries are removed immediately after enrichment. Stale entries (if AwardProgress
-        /// never fires for a node) don't survive beyond the save session — Subscribe(),
-        /// Unsubscribe(), and ResetForTesting() all clear the map before a new session can
-        /// route rewards through it.
-        /// </summary>
-        internal static readonly Dictionary<string, GameStateEvent> PendingMilestoneEventById
-            = new Dictionary<string, GameStateEvent>(StringComparer.Ordinal);
-
-        internal static void ClearPendingMilestoneEvents(string reason)
-        {
-            int cleared = PendingMilestoneEventById.Count;
-            if (cleared <= 0)
-                return;
-
-            PendingMilestoneEventById.Clear();
-            ParsekLog.Verbose("GameStateRecorder",
-                $"Cleared {cleared} pending milestone reward entr{(cleared == 1 ? "y" : "ies")} ({reason ?? "unspecified"})");
-        }
-
-        /// <summary>
-        /// Last completion UT seen per contract guid, used to debounce a stock contract
-        /// re-fire (Bug 1, fix-funds-economy-divergence §3.2). Stock KSP (or an external
-        /// interaction) can fire <c>Contract.onCompleted</c> N times for one contract within
-        /// a few milliseconds; without this debounce the recorder emits N
-        /// <see cref="GameStateEventType.ContractCompleted"/> events (not coalesced, since
-        /// they are not resource events), and the converter bakes N <c>ContractComplete</c>
-        /// ledger rows, multiplying the reward.
-        ///
-        /// Follows the <see cref="PendingMilestoneEventById"/> precedent exactly: a static
-        /// map cleared at every scene-entry / scene-exit and between tests via
-        /// <see cref="ClearContractCompletionDedup"/> (called from <see cref="Subscribe"/>,
-        /// <see cref="Unsubscribe"/>, and <see cref="ResetForTesting"/>). NO
-        /// <c>onGameSceneSwitchRequested</c> hook is added. Clearing the map on every scene
-        /// boundary keeps a legitimate re-completion in a different session / timeline from
-        /// being suppressed.
-        /// </summary>
-        internal static readonly Dictionary<string, double> lastContractCompletionUtByGuid
-            = new Dictionary<string, double>(StringComparer.Ordinal);
-
-        internal static void ClearContractCompletionDedup(string reason)
-        {
-            int cleared = lastContractCompletionUtByGuid.Count;
-            if (cleared <= 0)
-                return;
-
-            lastContractCompletionUtByGuid.Clear();
-            ParsekLog.Verbose("GameStateRecorder",
-                $"Cleared {cleared} contract-completion dedup entr{(cleared == 1 ? "y" : "ies")} ({reason ?? "unspecified"})");
-        }
-
-        /// <summary>
-        /// Pure debounce decision (Bug 1, fix-funds-economy-divergence §3.2). Returns true
-        /// (= duplicate, SUPPRESS) when <paramref name="contractGuid"/> is already present in
-        /// <paramref name="lastSeenByGuid"/> with a last-seen UT within
-        /// <paramref name="window"/> of <paramref name="ut"/>. Otherwise records
-        /// <c>lastSeenByGuid[contractGuid] = ut</c> and returns false (= first seen / genuine
-        /// re-completion, PROCEED). A null/empty guid is never treated as a duplicate (cannot
-        /// key it; let it through unchanged). Pure with respect to Unity (the caller passes
-        /// the static dict), so it is directly unit-testable.
-        /// </summary>
-        internal static bool IsDuplicateContractCompletion(
-            string contractGuid, double ut,
-            IDictionary<string, double> lastSeenByGuid, double window)
-        {
-            if (lastSeenByGuid == null || string.IsNullOrEmpty(contractGuid))
-                return false;
-
-            if (lastSeenByGuid.TryGetValue(contractGuid, out double lastSeen)
-                && Math.Abs(ut - lastSeen) <= window)
-            {
-                return true;
-            }
-
-            lastSeenByGuid[contractGuid] = ut;
-            return false;
-        }
-
-        /// <summary>
-        /// Called from the Harmony postfix on <c>ProgressNode.AwardProgress</c>. Looks
-        /// up the pending MilestoneAchieved event for the given node (by its qualified
-        /// id) and updates its detail string with the real funds/rep/sci values. If no
-        /// pending event is found (e.g., non-career mode where the subscriber didn't run),
-        /// this is a no-op. Also updates any matching ledger action that was already
-        /// written via <see cref="LedgerOrchestrator.OnKscSpending"/> — otherwise KSC
-        /// milestones would have zero rewards in the ledger.
-        /// Internal static for testability.
-        /// </summary>
-        internal static void EnrichPendingMilestoneRewards(
-            ProgressNode node, double funds, float rep, double sci)
-        {
-            if (node == null) return;
-            string milestoneId = QualifyMilestoneId(node);
-            if (string.IsNullOrEmpty(milestoneId))
-            {
-                ParsekLog.Warn("GameStateRecorder",
-                    "EnrichPendingMilestoneRewards: empty milestone id — cannot enrich");
-                return;
-            }
-            if (!PendingMilestoneEventById.TryGetValue(milestoneId, out var evt))
-            {
-                // #443: promoted from Verbose to Warn so the failure surfaces in default
-                // log settings. A miss here means the AwardProgress postfix ran without a
-                // matching OnProgressComplete-emitted pending event — usually a sign that
-                // either the dict was cleared between emit and enrich, or that the routing
-                // branch in ProgressRewardPatch.RoutePostfix sent us here incorrectly. Carry
-                // enough context (id, expected rewards, pending-map size) to triage.
-                ParsekLog.Warn("GameStateRecorder",
-                    $"EnrichPendingMilestoneRewards: no pending event for milestone id '{milestoneId}' " +
-                    $"(node.Id='{node.Id ?? "<null>"}', funds={funds:F0} rep={rep:F1} sci={sci:F1}, " +
-                    $"pendingMapSize={PendingMilestoneEventById.Count}) — skip");
-                return;
-            }
-
-            // Build the new detail string and patch the store's copy in place.
-            // GameStateEvent is a value type; the cached `evt` is a separate copy that
-            // we must also update for the test hooks and downstream observers.
-            string newDetail = BuildMilestoneDetail(funds, rep, sci);
-            bool storeUpdated = GameStateStore.UpdateEventDetail(evt, newDetail);
-            if (!storeUpdated)
-            {
-                // #443: defensive — UpdateEventDetail returning false used to be silent.
-                // Surface it so a future regression in the cached-vs-stored slot match
-                // is diagnosable from a default-level log. The trailing
-                // happy-path "Milestone enriched" Info is intentionally suppressed in this
-                // branch (logged inside the `else` below) so an operator never sees the
-                // contradicting "store had no matching event" Warn and "Milestone enriched"
-                // Info for the same call.
-                ParsekLog.Warn("GameStateRecorder",
-                    $"EnrichPendingMilestoneRewards: store had no matching event for " +
-                    $"'{evt.key}' ut={evt.ut:F1} tag='{evt.recordingId ?? ""}' — detail not patched");
-            }
-            evt.detail = newDetail;
-            PendingMilestoneEventById.Remove(milestoneId);
-
-            // If the event was already forwarded to the ledger via OnKscSpending, the
-            // ledger has a zero-reward MilestoneAchievement action we need to update.
-            // Find it by matching type + UT + milestoneId.
-            if (LedgerOrchestrator.IsInitialized)
-            {
-                var ledgerActions = Ledger.Actions;
-                for (int i = ledgerActions.Count - 1; i >= 0; i--)
-                {
-                    var a = ledgerActions[i];
-                    if (a.Type != GameActionType.MilestoneAchievement) continue;
-                    if (System.Math.Abs(a.UT - evt.ut) > 0.1) continue;
-                    if (a.MilestoneId != evt.key) continue;
-
-                    a.MilestoneFundsAwarded = (float)funds;
-                    a.MilestoneRepAwarded = rep;
-                    a.MilestoneScienceAwarded = (float)sci;
-                    ParsekLog.Verbose("GameStateRecorder",
-                        $"EnrichPendingMilestoneRewards: updated ledger action for '{evt.key}' " +
-                        $"funds={funds:F0} rep={rep:F1} sci={sci:F1}");
-                    break;
-                }
-            }
-
-            // #443 review: only print the happy-path Info when the store side actually
-            // got patched. The earlier shape printed it unconditionally, so on a
-            // store-miss the operator saw a contradicting Warn + Info pair for the same
-            // call. On the miss branch we leave the Warn above as the sole signal and
-            // demote the in-memory-only success line to Verbose for full-fidelity replay.
-            if (storeUpdated)
-            {
-                ParsekLog.Info("GameStateRecorder",
-                    $"Milestone enriched: '{evt.key}' funds={funds:F0} rep={rep:F1} sci={sci:F1}");
-            }
-            else
-            {
-                ParsekLog.Verbose("GameStateRecorder",
-                    $"Milestone enriched (ledger only, store skipped): '{evt.key}' " +
-                    $"funds={funds:F0} rep={rep:F1} sci={sci:F1}");
-            }
-        }
-
-        /// <summary>
-        /// #442: emits a fully-populated <see cref="GameStateEventType.MilestoneAchieved"/>
-        /// event for progress nodes whose subclass calls <c>AwardProgress</c> directly
-        /// without going through <c>OnProgressComplete</c> (e.g. <c>RecordsSpeed</c>,
-        /// <c>RecordsAltitude</c>, <c>RecordsDistance</c>, <c>RecordsDepth</c>). Without
-        /// this path the world-record FundsChanged(Progression) deltas are dropped wholesale
-        /// by <see cref="GameStateEventConverter"/>'s drop-rule and the rewards never reach
-        /// the ledger.
-        ///
-        /// Called from the Harmony postfix on <c>ProgressNode.AwardProgress</c> only when
-        /// no entry exists in <see cref="PendingMilestoneEventById"/> — i.e. when the
-        /// usual two-phase emit-then-enrich path did not run. Bypasses the enrichment-map
-        /// indirection entirely: rewards arrive as method parameters and are baked into
-        /// the event detail directly.
-        ///
-        /// <paramref name="ut"/> is supplied by the caller — production passes
-        /// <see cref="Planetarium.GetUniversalTime"/> from the patch site (which has Unity
-        /// statics live), unit tests pass a literal so they don't NRE on Planetarium.
-        /// Internal static for testability.
-        /// </summary>
-        internal static void EmitStandaloneProgressReward(
-            ProgressNode node, double funds, float rep, double sci, double ut)
-        {
-            if (IsReplayingActions)
-            {
-                ParsekLog.Verbose("GameStateRecorder",
-                    "EmitStandaloneProgressReward: suppressed during action replay");
-                return;
-            }
-            if (node == null)
-            {
-                ParsekLog.Verbose("GameStateRecorder",
-                    "EmitStandaloneProgressReward: null node — skipped");
-                return;
-            }
-
-            string milestoneId = QualifyMilestoneId(node);
-            if (string.IsNullOrEmpty(milestoneId))
-            {
-                ParsekLog.Verbose("GameStateRecorder",
-                    "EmitStandaloneProgressReward: empty node Id — skipped");
-                return;
-            }
-
-            var evt = new GameStateEvent
-            {
-                ut = ut,
-                eventType = GameStateEventType.MilestoneAchieved,
-                key = milestoneId,
-                detail = BuildMilestoneDetail(funds, rep, sci)
-            };
-            Emit(ref evt, "MilestoneAchievedStandalone");
-
-            ParsekLog.Info("GameStateRecorder",
-                $"Game state: MilestoneAchieved (standalone) '{milestoneId}' " +
-                $"funds={funds:F0} rep={rep:F1} sci={sci:F1}");
-
-            // Mirror the OnProgressComplete KSC-forwarding path so world-record rewards
-            // earned outside flight (rare, but possible — e.g. RecordsAltitude on a sub-
-            // orbital craft already reverted) still reach the ledger immediately.
-            // #553 follow-up: gate on ShouldForwardDirectLedgerEvent so untagged
-            // pre-recording FLIGHT standalone milestone awards reach the ledger too.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        /// <summary>
-        /// Reflection helper: extracts the private "body" CelestialBody field from
-        /// body-specific ProgressNode subclasses (CelestialBodyLanding, CelestialBodyOrbit,
-        /// CelestialBodyFlyby, etc.). Returns null for top-level nodes that have no body.
-        /// Internal static for testability.
-        /// </summary>
-        internal static CelestialBody ExtractNodeBody(ProgressNode node)
-        {
-            if (node == null) return null;
-            var bodyField = node.GetType().GetField("body",
-                BindingFlags.NonPublic | BindingFlags.Instance);
-            if (bodyField == null) return null;
-            return bodyField.GetValue(node) as CelestialBody;
-        }
-
-        /// <summary>
-        /// Pure: creates a MilestoneAchieved GameStateEvent from the given parameters.
-        /// Extracted for testability.
-        /// </summary>
-        internal static GameStateEvent CreateMilestoneEvent(string milestoneId, double ut)
-        {
-            return new GameStateEvent
-            {
-                ut = ut,
-                eventType = GameStateEventType.MilestoneAchieved,
-                key = milestoneId ?? "",
-                detail = ""
-            };
-        }
-
-        /// <summary>
-        /// Path-qualifies body-specific milestone IDs to avoid ambiguity.
-        /// Body-specific achievement classes (CelestialBodyLanding, CelestialBodyOrbit, etc.)
-        /// all have a private CelestialBody body field. The bare node.Id (e.g. "Landing")
-        /// is non-unique across bodies, so we qualify as "Mun/Landing", "Kerbin/Landing", etc.
-        ///
-        /// Top-level nodes (FirstLaunch, ReachSpace, RecordsAltitude, etc.) have unique IDs
-        /// and are returned as-is.
-        ///
-        /// Internal static for testability.
-        /// </summary>
-        internal static string QualifyMilestoneId(ProgressNode node)
-        {
-            if (node == null) return "";
-
-            string bareId = node.Id ?? "";
-            if (string.IsNullOrEmpty(bareId)) return "";
-
-            // Try reflection to get the private "body" field present on all
-            // CelestialBody* achievement subclasses (Landing, Orbit, Flyby, etc.)
-            var bodyField = node.GetType().GetField("body",
-                BindingFlags.NonPublic | BindingFlags.Instance);
-            if (bodyField != null)
-            {
-                var celestialBody = bodyField.GetValue(node) as CelestialBody;
-                if (celestialBody != null)
-                {
-                    string qualified = celestialBody.name + "/" + bareId;
-                    ParsekLog.Verbose("GameStateRecorder",
-                        $"QualifyMilestoneId: qualified '{bareId}' -> '{qualified}' (body={celestialBody.name})");
-                    return qualified;
-                }
-            }
-
-            // No body field — top-level node with unique ID (FirstLaunch, ReachSpace, etc.)
-            return bareId;
-        }
-
-        #endregion
-
-        #region Strategy Lifecycle (#439 Phase A)
-
-        /// <summary>
-        /// Called from the Harmony postfix on <c>Strategies.Strategy.Activate</c> when
-        /// <c>__result == true</c> (stock returns false on a CanBeActivated miss). Emits a
-        /// <see cref="GameStateEventType.StrategyActivated"/> event carrying the strategy's
-        /// configured setup costs; downstream <see cref="GameStateEventConverter"/> parses
-        /// the detail and the ledger's <see cref="LedgerOrchestrator.ClassifyAction"/>
-        /// reconciles the funds setup cost against the paired
-        /// <c>FundsChanged(StrategySetup)</c> event KSP fires inside <c>Activate()</c>.
-        ///
-        /// Respects <see cref="IsReplayingActions"/> so recalculation replays stay silent.
-        /// Internal static for testability — the Harmony postfix is the only production
-        /// call site.
-        /// </summary>
-        internal static void OnStrategyActivated(Strategies.Strategy strategy)
-        {
-            if (IsReplayingActions)
-            {
-                ParsekLog.Verbose("GameStateRecorder",
-                    "OnStrategyActivated: suppressed during action replay");
-                return;
-            }
-            if (strategy == null)
-            {
-                ParsekLog.Verbose("GameStateRecorder",
-                    "OnStrategyActivated: null strategy - skipped");
-                return;
-            }
-            if (strategy.Config == null)
-            {
-                ParsekLog.Warn("GameStateRecorder",
-                    "OnStrategyActivated: strategy.Config is null - skipped");
-                return;
-            }
-
-            string key = strategy.Config.Name ?? "";
-            string title = strategy.Title ?? "";
-            string dept = strategy.DepartmentName ?? "";
-            float factor = strategy.Factor;
-            float setupFunds = strategy.InitialCostFunds;
-            float setupSci = strategy.InitialCostScience;
-            float setupRep = strategy.InitialCostReputation;
-            StrategyResource sourceResource;
-            StrategyResource targetResource;
-            if (!TryExtractStrategyResourceFlow(strategy, out sourceResource, out targetResource))
-            {
-                sourceResource = StrategyResource.Funds;
-                targetResource = StrategyResource.Funds;
-            }
-
-            var evt = new GameStateEvent
-            {
-                ut = Planetarium.GetUniversalTime(),
-                eventType = GameStateEventType.StrategyActivated,
-                key = key,
-                detail = BuildStrategyActivateDetail(title, dept, factor,
-                    setupFunds, setupSci, setupRep,
-                    sourceResource: sourceResource, targetResource: targetResource)
-            };
-            Emit(ref evt, "StrategyActivated");
-
-            ParsekLog.Info("GameStateRecorder",
-                $"Game state: StrategyActivated '{key}' title='{title}' dept='{dept}' " +
-                $"factor={factor.ToString("F2", System.Globalization.CultureInfo.InvariantCulture)} " +
-                $"setupFunds={setupFunds.ToString("F0", System.Globalization.CultureInfo.InvariantCulture)} " +
-                $"setupSci={setupSci.ToString("F1", System.Globalization.CultureInfo.InvariantCulture)} " +
-                $"setupRep={setupRep.ToString("F1", System.Globalization.CultureInfo.InvariantCulture)}");
-
-            // KSC-side forwarding mirror: StrategyActivate actions with a non-zero setup
-            // cost need to land on the ledger immediately when the player activates from
-            // the Administration building (i.e. outside flight). The commit-time
-            // ConvertEvents path will also pick up flight-scope strategies. The
-            // ShouldForwardDirectLedgerEvent gate suppresses tagged teardown events
-            // while still letting untagged pre-recording FLIGHT events (no live
-            // recorder) reach the ledger.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        /// <summary>
-        /// Called from the Harmony postfix on <c>Strategies.Strategy.Deactivate</c> when
-        /// <c>__result == true</c>. Emits <see cref="GameStateEventType.StrategyDeactivated"/>.
-        /// Stock Deactivate has no resource cost, so the classifier maps this to
-        /// <see cref="KscActionExpectationClassifier.KscReconcileClass.NoResourceImpact"/>.
-        /// Internal static for testability.
-        /// </summary>
-        internal static void OnStrategyDeactivated(Strategies.Strategy strategy)
-        {
-            if (IsReplayingActions)
-            {
-                ParsekLog.Verbose("GameStateRecorder",
-                    "OnStrategyDeactivated: suppressed during action replay");
-                return;
-            }
-            if (strategy == null)
-            {
-                ParsekLog.Verbose("GameStateRecorder",
-                    "OnStrategyDeactivated: null strategy - skipped");
-                return;
-            }
-            if (strategy.Config == null)
-            {
-                ParsekLog.Warn("GameStateRecorder",
-                    "OnStrategyDeactivated: strategy.Config is null - skipped");
-                return;
-            }
-
-            string key = strategy.Config.Name ?? "";
-            string title = strategy.Title ?? "";
-            string dept = strategy.DepartmentName ?? "";
-            float factor = strategy.Factor;
-            double now = Planetarium.GetUniversalTime();
-            double activeDurationSec = now - strategy.DateActivated;
-            if (activeDurationSec < 0) activeDurationSec = 0;
-
-            var evt = new GameStateEvent
-            {
-                ut = now,
-                eventType = GameStateEventType.StrategyDeactivated,
-                key = key,
-                detail = BuildStrategyDeactivateDetail(title, dept, factor, activeDurationSec)
-            };
-            Emit(ref evt, "StrategyDeactivated");
-
-            ParsekLog.Info("GameStateRecorder",
-                $"Game state: StrategyDeactivated '{key}' title='{title}' dept='{dept}' " +
-                $"factor={factor.ToString("F2", System.Globalization.CultureInfo.InvariantCulture)} " +
-                $"activeDurationSec={activeDurationSec.ToString("F1", System.Globalization.CultureInfo.InvariantCulture)}");
-
-            // Mirror the KSC-forwarding path even though StrategyDeactivate is a
-            // NoResourceImpact action — the ledger still needs the StrategyDeactivate row
-            // so StrategiesModule can pair activate/deactivate during the walk.
-            // #553 follow-up: gate on ShouldForwardDirectLedgerEvent so untagged
-            // pre-recording FLIGHT strategy deactivations reach the ledger too.
-            if (ShouldForwardDirectLedgerEvent(evt.recordingId, HasLiveRecorder()))
-                LedgerOrchestrator.OnKscSpending(evt);
-        }
-
-        /// <summary>
-        /// Pure: builds the semicolon-separated detail string for a StrategyActivated
-        /// event. Format:
-        /// <c>title=&lt;t&gt;;dept=&lt;d&gt;;factor=&lt;f&gt;;source=&lt;src&gt;;target=&lt;tgt&gt;;setupFunds=&lt;sf&gt;;setupSci=&lt;ss&gt;;setupRep=&lt;sr&gt;</c>.
-        /// Numerics serialized with InvariantCulture "R" (round-trip) so comma-locale
-        /// hosts do not break the converter-side parse.
-        /// Internal static for direct unit testing without a KSP runtime.
-        /// </summary>
-        internal static string BuildStrategyActivateDetail(
-            string title, string dept, float factor,
-            float setupFunds, float setupSci, float setupRep,
-            StrategyResource sourceResource = StrategyResource.Funds,
-            StrategyResource targetResource = StrategyResource.Funds)
-        {
-            var ic = System.Globalization.CultureInfo.InvariantCulture;
-            return $"title={title ?? ""};" +
-                   $"dept={dept ?? ""};" +
-                   $"factor={factor.ToString("R", ic)};" +
-                   $"source={sourceResource};" +
-                   $"target={targetResource};" +
-                   $"setupFunds={setupFunds.ToString("R", ic)};" +
-                   $"setupSci={setupSci.ToString("R", ic)};" +
-                   $"setupRep={setupRep.ToString("R", ic)}";
-        }
-
-        /// <summary>
-        /// Pure: builds the semicolon-separated detail string for a StrategyDeactivated
-        /// event. Format: <c>title=&lt;t&gt;;dept=&lt;d&gt;;factor=&lt;f&gt;;activeDurationSec=&lt;d&gt;</c>.
-        /// Internal static for direct unit testing.
-        /// </summary>
-        internal static string BuildStrategyDeactivateDetail(
-            string title, string dept, float factor, double activeDurationSec)
-        {
-            var ic = System.Globalization.CultureInfo.InvariantCulture;
-            return $"title={title ?? ""};" +
-                   $"dept={dept ?? ""};" +
-                   $"factor={factor.ToString("R", ic)};" +
-                   $"activeDurationSec={activeDurationSec.ToString("R", ic)}";
-        }
-
-        internal static bool TryExtractStrategyResourceFlow(
-            Strategies.Strategy strategy,
-            out StrategyResource sourceResource,
-            out StrategyResource targetResource)
-        {
-            sourceResource = StrategyResource.Funds;
-            targetResource = StrategyResource.Funds;
-
-            if (strategy == null || strategy.Effects == null)
-                return false;
-
-            for (int i = 0; i < strategy.Effects.Count; i++)
-            {
-                object effect = strategy.Effects[i];
-                if (effect == null)
-                    continue;
-
-                if (TryExtractCurrencyConverterFlow(effect, out sourceResource, out targetResource))
-                    return true;
-
-                if (TryExtractCurrencyOperationFlow(effect, out sourceResource))
-                {
-                    targetResource = sourceResource;
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        private static bool TryExtractCurrencyConverterFlow(
-            object effect,
-            out StrategyResource sourceResource,
-            out StrategyResource targetResource)
-        {
-            sourceResource = StrategyResource.Funds;
-            targetResource = StrategyResource.Funds;
-
-            Type effectType = effect.GetType();
-            if (!string.Equals(effectType.FullName, "Strategies.Effects.CurrencyConverter",
-                    StringComparison.Ordinal))
-                return false;
-
-            var inputField = effectType.GetField("input",
-                BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
-            var outputField = effectType.GetField("output",
-                BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
-            if (inputField == null || outputField == null)
-                return false;
-
-            return TryParseStrategyCurrency(inputField.GetValue(effect), out sourceResource)
-                && TryParseStrategyCurrency(outputField.GetValue(effect), out targetResource);
-        }
-
-        private static bool TryExtractCurrencyOperationFlow(
-            object effect,
-            out StrategyResource resource)
-        {
-            resource = StrategyResource.Funds;
-
-            Type effectType = effect.GetType();
-            if (!string.Equals(effectType.FullName, "Strategies.Effects.CurrencyOperation",
-                    StringComparison.Ordinal))
-                return false;
-
-            var currencyField = effectType.GetField("currency",
-                BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
-            return currencyField != null
-                && TryParseStrategyCurrency(currencyField.GetValue(effect), out resource);
-        }
-
-        private static bool TryParseStrategyCurrency(
-            object rawCurrency,
-            out StrategyResource resource)
-        {
-            resource = StrategyResource.Funds;
-            if (rawCurrency == null)
-                return false;
-
-            int value;
-            try
-            {
-                value = Convert.ToInt32(rawCurrency, System.Globalization.CultureInfo.InvariantCulture);
-            }
-            catch
-            {
-                return false;
-            }
-
-            switch (value)
-            {
-                case 0:
-                    resource = StrategyResource.Funds;
-                    return true;
-                case 1:
-                    resource = StrategyResource.Science;
-                    return true;
-                case 2:
-                    resource = StrategyResource.Reputation;
-                    return true;
-                default:
-                    return false;
-            }
         }
 
         #endregion
