+++ scripts/ers-els-audit-allowlist.txt
@@ -36,6 +36,14 @@ Source/Parsek/KerbalsModule.cs
 # --- Internal game-state-event suppression check ---
 Source/Parsek/GameStateRecorder.cs
 
+# GameStateRecorder.Handlers.cs is the partial-class split holding the
+# contract/tech/crew/progress-milestone/strategy event-handler regions.
+# EnrichPendingMilestoneRewards (moved verbatim from GameStateRecorder.cs)
+# scans Ledger.Actions directly to update the already-forwarded
+# zero-reward MilestoneAchievement action in place — same trust scope as
+# GameStateRecorder.cs above.
+Source/Parsek/GameStateRecorder.Handlers.cs
+
 # --- Tests (unit + in-game) ---
 Source/Parsek.Tests/
 Source/Parsek/InGameTests/
